CS236/Homeworks/HW1/GPT2/requirements.txt
matplotlib
numpy
pyyaml
torch
tqdm
regex